+++ packages/jest-haste-map/src/index.ts
@@ -29,6 +29,7 @@ import normalizePathSep from './lib/normalizePathSep';
 import watchmanCrawl from './crawlers/watchman';
 // @ts-ignore: not converted to TypeScript - it's a fork: https://github.com/facebook/jest/pull/5387
 import WatchmanWatcher from './lib/WatchmanWatcher';
+import FSEventsWatcher from './lib/FSEventsWatcher';
 import * as fastPath from './lib/fast_path';
 import {
   ChangeEvent,
@@ -795,10 +796,14 @@ class HasteMap extends EventEmitter {
     this._options.throwOnModuleCollision = false;
     this._options.retainAllFiles = true;
 
+    // WatchmanWatcher > FSEventsWatcher > sane.NodeWatcher
     const Watcher: sane.Watcher =
       canUseWatchman && this._options.useWatchman
         ? WatchmanWatcher
+        : FSEventsWatcher.isSupported()
+        ? FSEventsWatcher
         : sane.NodeWatcher;
+
     const extensions = this._options.extensions;
     const ignorePattern = this._options.ignorePattern;
     const rootDir = this._options.rootDir;